% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw04dwjkate-package.R
\docType{package}
\name{hw04dwjkate-package}
\alias{hw04dwjkate}
\alias{hw04dwjkate-package}
\title{hw04dwjkate: STAT 613 Homework: Creating a Package}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Dawoon (Kate) Jung \email{dj1997a@student.american.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
