% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursivesequence.R
\name{myseq_n}
\alias{myseq_n}
\title{Recursive Sequence}
\usage{
myseq_n(x, n)
}
\arguments{
\item{x}{Vector containing three numeric elements}

\item{n}{Number of elements of the sequence}
}
\value{
nth element of the sequence
}
\description{
Recursive Sequence
}
\examples{
myseq_n(x = c(2, 4, 3), n = 7)
}
