% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceplot.R
\name{seq_plot}
\alias{seq_plot}
\title{Sequence Plot}
\usage{
seq_plot(df)
}
\arguments{
\item{df}{A Dataframe with four columns - all numbers}
}
\value{
A plot showing the result of the myseq_n() function
}
\description{
Sequence Plot
}
\examples{
my_data <- tibble::tribble(
~x, ~y, ~z, ~n,
2,4,3,3,
2,4,3,4,
2,4,3,5,
2,4,3,6,
2,4,3,7,
2,4,3,8,
2,4,3,9,
2,4,3,10,
2,4,3,12)
seq_plot(my_data)
}
